refactor(test): consolidate duplicated HomeGuard into 3 shared helpers
Same struct/impl was hand-duplicated across 8 files. Down to 3 shared
copies split along crate-visibility boundaries: src/session/test_support.rs
(library crate), tests/e2e/harness.rs (the e2e test binary), and the new
tests/home_isolation/mod.rs (standalone tests/*.rs binaries that don't
also need tests/common's port helpers).
