+++ tests/e2e/filewatch_tui_burst_reload.rs
@@ -13,54 +13,7 @@ use agent_of_empires::file_watch::FileWatchService;
 use agent_of_empires::session::{Instance, Storage};
 use serial_test::serial;
 
-use crate::harness::{require_tmux, TuiTestHarness};
-
-/// RAII guard: points `HOME`/`XDG_CONFIG_HOME` at the harness's tempdir
-/// for the test process and restores the prior values on `Drop`.
-/// `#[serial]` on every caller linearizes this against other tests in
-/// the binary; without the restore, a later test could inherit this
-/// test's (by-then-dropped) tempdir path.
-#[must_use = "HomeGuard restores env vars on Drop; bind it, don't discard it, or isolation ends immediately"]
-struct HomeGuard {
-    prev_home: Option<std::ffi::OsString>,
-    prev_xdg: Option<std::ffi::OsString>,
-}
-
-impl HomeGuard {
-    /// Snapshots the current `HOME`/`XDG_CONFIG_HOME` before overriding them,
-    /// so `Drop` can restore the caller's real environment.
-    fn new(home: &std::path::Path) -> Self {
-        let prev_home = std::env::var_os("HOME");
-        let prev_xdg = std::env::var_os("XDG_CONFIG_HOME");
-        // SAFETY: env mutation; #[serial] linearizes this against every
-        // other #[serial] test in the binary, so no concurrent
-        // reader/writer exists.
-        unsafe { std::env::set_var("HOME", home) };
-        unsafe { std::env::set_var("XDG_CONFIG_HOME", home.join(".config")) };
-        Self {
-            prev_home,
-            prev_xdg,
-        }
-    }
-}
-
-impl Drop for HomeGuard {
-    fn drop(&mut self) {
-        /// Restores `key` to its prior value, or removes it if it was
-        /// previously unset.
-        fn restore_or_remove(key: &str, prev: Option<std::ffi::OsString>) {
-            // SAFETY: same invariant as HomeGuard::new; #[serial] guards this.
-            unsafe {
-                match prev {
-                    Some(v) => std::env::set_var(key, v),
-                    None => std::env::remove_var(key),
-                }
-            }
-        }
-        restore_or_remove("HOME", self.prev_home.take());
-        restore_or_remove("XDG_CONFIG_HOME", self.prev_xdg.take());
-    }
-}
+use crate::harness::{require_tmux, HomeGuard, TuiTestHarness};
 
 #[test]
 #[serial]
